[resource-manager] Move `pkg/resourcemanager/controller/node` to `pkg/resourcemanager/controller/node/criticalcomponents`
We will introduce another controller for nodes, so let's prepare the folder structure
