+++ pkg/resourcemanager/controller/node/add.go
@@ -5,68 +5,24 @@
 package node
 
 import (
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/utils/ptr"
-	"sigs.k8s.io/controller-runtime/pkg/builder"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/cluster"
-	"sigs.k8s.io/controller-runtime/pkg/controller"
-	"sigs.k8s.io/controller-runtime/pkg/handler"
-	"sigs.k8s.io/controller-runtime/pkg/manager"
-	"sigs.k8s.io/controller-runtime/pkg/predicate"
-	"sigs.k8s.io/controller-runtime/pkg/source"
+	"fmt"
 
-	v1beta1constants "github.com/gardener/gardener/pkg/apis/core/v1beta1/constants"
-	predicateutils "github.com/gardener/gardener/pkg/controllerutils/predicate"
+	"sigs.k8s.io/controller-runtime/pkg/cluster"
+	"sigs.k8s.io/controller-runtime/pkg/manager"
+
+	"github.com/gardener/gardener/pkg/resourcemanager/apis/config"
+	"github.com/gardener/gardener/pkg/resourcemanager/controller/node/criticalcomponents"
 )
 
-// ControllerName is the name of the controller.
-const ControllerName = "node"
-
-// AddToManager adds Reconciler to the given manager.
-func (r *Reconciler) AddToManager(mgr manager.Manager, targetCluster cluster.Cluster) error {
-	if r.TargetClient == nil {
-		r.TargetClient = targetCluster.GetClient()
-	}
-	if r.Recorder == nil {
-		r.Recorder = targetCluster.GetEventRecorderFor("gardener-" + ControllerName + "-controller") // node-controller is ambiguous
-	}
-
-	return builder.
-		ControllerManagedBy(mgr).
-		Named(ControllerName).
-		WithOptions(controller.Options{
-			MaxConcurrentReconciles: ptr.Deref(r.Config.ConcurrentSyncs, 0),
-		}).
-		WatchesRawSource(
-			source.Kind(targetCluster.GetCache(), &corev1.Node{}),
-			&handler.EnqueueRequestForObject{},
-			builder.WithPredicates(r.NodePredicate()),
-		).
-		Complete(r)
-}
-
-// NodePredicate returns a predicate that filters for Node objects that are created with the taint.
-func (r *Reconciler) NodePredicate() predicate.Predicate {
-	return predicate.And(
-		predicateutils.ForEventTypes(predicateutils.Create),
-		predicate.NewPredicateFuncs(func(obj client.Object) bool {
-			return NodeHasCriticalComponentsNotReadyTaint(obj)
-		}),
-	)
-}
-
-// NodeHasCriticalComponentsNotReadyTaint returns true if the given Node has the taint that this controller manages.
-func NodeHasCriticalComponentsNotReadyTaint(obj client.Object) bool {
-	node, ok := obj.(*corev1.Node)
-	if !ok {
-		return false
-	}
-
-	for _, taint := range node.Spec.Taints {
-		if taint.Key == v1beta1constants.TaintNodeCriticalComponentsNotReady {
-			return true
+// AddToManager adds all node controllers to the given manager.
+func AddToManager(mgr manager.Manager, targetCluster cluster.Cluster, cfg config.ResourceManagerConfiguration) error {
+	if cfg.Controllers.Node.Enabled {
+		if err := (&criticalcomponents.Reconciler{
+			Config: cfg.Controllers.Node,
+		}).AddToManager(mgr, targetCluster); err != nil {
+			return fmt.Errorf("failed adding node-critical-components controller: %w", err)
 		}
 	}
-	return false
+
+	return nil
 }
