+++ pkg/resourcemanager/controller/node/criticalcomponents/reconciler_test.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package node_test
+package criticalcomponents_test
 
 import (
 	"context"
@@ -28,7 +28,7 @@ import (
 	logzap "sigs.k8s.io/controller-runtime/pkg/log/zap"
 
 	"github.com/gardener/gardener/pkg/logger"
-	. "github.com/gardener/gardener/pkg/resourcemanager/controller/node"
+	. "github.com/gardener/gardener/pkg/resourcemanager/controller/node/criticalcomponents"
 	"github.com/gardener/gardener/pkg/utils"
 	"github.com/gardener/gardener/pkg/utils/test"
 	mockclient "github.com/gardener/gardener/third_party/mock/controller-runtime/client"
