+++ skaffold-operator.yaml
@@ -269,7 +269,8 @@ build:
             - pkg/resourcemanager/controller/managedresource
             - pkg/resourcemanager/controller/networkpolicy
             - pkg/resourcemanager/controller/node
-            - pkg/resourcemanager/controller/node/helper
+            - pkg/resourcemanager/controller/node/criticalcomponents
+            - pkg/resourcemanager/controller/node/criticalcomponents/helper
             - pkg/resourcemanager/controller/tokeninvalidator
             - pkg/resourcemanager/predicate
             - pkg/resourcemanager/webhook
